Move StringOption to its own source file
